# Rewrite more Map-Reduces as Renderers to save garbage