--- a/render/container.go
+++ b/render/container.go
@@ -82,7 +82,7 @@ func (c connectionJoin) Render(rpt report.Report) Nodes {
 		// Nodes without a hostid may be pseudo nodes - if so, pass through to result
 		if _, ok := m.Latest.Lookup(report.HostNodeID); !ok {
 			if id, ok := externalNodeID(m, addr, local); ok {
-				ret.addToResults(m, id, newPseudoNode)
+				ret.addChild(m, id, newPseudoNode)
 				continue
 			}
 		}
@@ -94,7 +94,7 @@ func (c connectionJoin) Render(rpt report.Report) Nodes {
 			id, found = ipNodes[report.MakeScopedEndpointNodeID(scope, addr, port)]
 		}
 		if found && id != "" { // not one we blanked out earlier
-			ret.addToResults(m, id, func(id string) report.Node {
+			ret.addChild(m, id, func(id string) report.Node {
 				return inputNodes.Nodes[id]
 			})
 		}