`MapX2Host` and to a lesser extent `MapProcess2Name` have the property that they map many nodes onto fewer nodes, so looping and using the machinery in `joinResults` is more effective than creating many copies of the same node and merging.

Benchmarks before (c5bdebdffebfa7cc0839770a3bf70429260f1bc9):
```
BenchmarkTopologyList-2         	       2	 603197223 ns/op	112161732 B/op	 1253503 allocs/op
BenchmarkTopologyHosts-2        	       3	 385415232 ns/op	82957173 B/op	 1015204 allocs/op
BenchmarkTopologyContainers-2   	      10	 192766565 ns/op	33145659 B/op	  407698 allocs/op
```

after this PR:
```
BenchmarkTopologyList-2         	       3	 476705872 ns/op	101366640 B/op	 1109042 allocs/op
BenchmarkTopologyHosts-2        	       3	 373288182 ns/op	76540733 B/op	  933301 allocs/op
BenchmarkTopologyContainers-2   	      10	 183615683 ns/op	33132417 B/op	  407672 allocs/op
```