@rade 2017-11-22
With more-renderers-b I am seeing a similar improvements in allocations as above, but no improvements in runtime:

master
```
BenchmarkTopologyList-4         	       2	 634468221 ns/op	107948256 B/op	 1056511 allocs/op
BenchmarkTopologyHosts-4        	       2	 553097064 ns/op	78231540 B/op	  812637 allocs/op
BenchmarkTopologyContainers-4   	       5	 235253550 ns/op	30075289 B/op	  302554 allocs/op
```
branch
```
BenchmarkTopologyList-4         	       2	 638986932 ns/op	97484960 B/op	  915473 allocs/op
BenchmarkTopologyHosts-4        	       2	 545031910 ns/op	71767348 B/op	  730887 allocs/op
BenchmarkTopologyContainers-4   	       5	 225194107 ns/op	30069329 B/op	  302520 allocs/op
```
(each line is the best result of three runs)