@rade 2017-11-22
I've re-run the benchmarks with `cpu=1`, and now I _do_ see a >5% runtime improvement in `BenchmarkTopologyList`.